Extract inner classes out of FeatureDataSource to improve readability and modularity
Preparation for some refactoring related to TileDB integration

Extracted classes are package-protected for now, since they are not intended
for direct use outside of the engine package.
