+++ src/main/java/org/broadinstitute/hellbender/engine/FeatureDataSource.java
@@ -111,219 +111,6 @@ public final class FeatureDataSource<T extends Feature> implements GATKDataSourc
     public static final int DEFAULT_QUERY_LOOKAHEAD_BASES = 1000;
 
     /**
-     * FeatureCache: helper class to manage the cache of Feature records used during query operations
-     * initiated via {@link #query(SimpleInterval)} and/or {@link #queryAndPrefetch(SimpleInterval)}.
-     *
-     * Strategy is to pre-fetch a large number of records AFTER each query interval that produces
-     * a cache miss. This optimizes for the use case of intervals with gradually increasing start
-     * positions, as many subsequent queries will find their records wholly contained in the cache
-     * before we have another cache miss. Performance will be poor for random/non-localized access
-     * patterns, or intervals with decreasing start positions.
-     *
-     * Usage:
-     * -Test whether each query interval is a cache hit via {@link #cacheHit(SimpleInterval)}
-     *
-     * -If it is a cache hit, trim the cache to the start position of the interval (discarding records that
-     *  end before the start of the new interval) via {@link #trimToNewStartPosition(int)}, then retrieve
-     *  records up to the desired endpoint using {@link #getCachedFeaturesUpToStopPosition(int)}.
-     *
-     * -If it is a cache miss, reset the cache using {@link #fill(Iterator, SimpleInterval)}, pre-fetching
-     *  a large number of records after the query interval in addition to those actually requested.
-     *
-     * @param <CACHED_FEATURE> Type of Feature record we are caching
-     */
-    protected static class FeatureCache<CACHED_FEATURE extends Feature> {
-
-        /**
-         * Our cache of Features, optimized for insertion/removal at both ends.
-         */
-        private final Deque<CACHED_FEATURE> cache;
-
-        /**
-         * Our cache currently contains Feature records overlapping this interval
-         */
-        private SimpleInterval cachedInterval;
-
-        /**
-         * Number of times we called {@link #cacheHit(SimpleInterval)} and it returned true
-         */
-        private int numCacheHits = 0;
-
-        /**
-         * Number of times we called {@link #cacheHit(SimpleInterval)} and it returned false
-         */
-        private int numCacheMisses = 0;
-
-        /**
-         * Initial capacity of our cache (will grow by doubling if needed)
-         */
-        private static final int INITIAL_CAPACITY = 1024;
-
-        /**
-         * When we trim our cache to a new start position, this is the maximum number of
-         * Features we expect to need to place into temporary storage for the duration of
-         * the trim operation. Performance only suffers slightly if our estimate is wrong.
-         */
-        private static final int EXPECTED_MAX_OVERLAPPING_FEATURES_DURING_CACHE_TRIM = 128;
-
-        /**
-         * Create an initially-empty FeatureCache with default initial capacity
-         */
-        public FeatureCache() {
-            cache = new ArrayDeque<>(INITIAL_CAPACITY);
-        }
-
-        /**
-         * Get the name of the contig on which the Features in our cache are located
-         *
-         * @return the name of the contig on which the Features in our cache are located
-         */
-        public String getContig() {
-            return cachedInterval.getContig();
-        }
-
-        /**
-         * Get the start position of the interval that all Features in our cache overlap
-         *
-         * @return the start position of the interval that all Features in our cache overlap
-         */
-        public int getCacheStart() {
-            return cachedInterval.getStart();
-        }
-
-        /**
-         * Get the stop position of the interval that all Features in our cache overlap
-         *
-         * @return the stop position of the interval that all Features in our cache overlap
-         */
-        public int getCacheEnd() {
-            return cachedInterval.getEnd();
-        }
-
-        /**
-         * Does our cache currently contain no Features?
-         *
-         * @return true if our cache contains no Features, otherwise false
-         */
-        public boolean isEmpty() {
-            return cache.isEmpty();
-        }
-
-        /**
-         * @return Number of times we called {@link #cacheHit(SimpleInterval)} and it returned true
-         */
-        public int getNumCacheHits() {
-            return numCacheHits;
-        }
-
-        /**
-         * @return Number of times we called {@link #cacheHit(SimpleInterval)} and it returned false
-         */
-        public int getNumCacheMisses() {
-            return numCacheMisses;
-        }
-
-        /**
-         * Clear our cache and fill it with the records from the provided iterator, preserving their
-         * relative ordering, and update our contig/start/stop to reflect the new interval that all
-         * records in our cache overlap.
-         *
-         * Typically each fill operation should involve significant lookahead beyond the region
-         * requested so that future queries will be cache hits.
-         *
-         * @param featureIter iterator from which to pull Features with which to populate our cache
-         *                    (replacing existing cache contents)
-         * @param interval all Features from featureIter overlap this interval
-         */
-        public void fill( final Iterator<CACHED_FEATURE> featureIter, final SimpleInterval interval ) {
-            cache.clear();
-            while ( featureIter.hasNext() ) {
-                cache.add(featureIter.next());
-            }
-
-            cachedInterval = interval;
-        }
-
-        /**
-         * Determines whether all records overlapping the provided interval are already contained in our cache.
-         *
-         * @param interval the interval to check against the contents of our cache
-         * @return true if all records overlapping the provided interval are already contained in our cache, otherwise false
-         */
-        public boolean cacheHit( final SimpleInterval interval ) {
-            final boolean cacheHit = cachedInterval != null && cachedInterval.contains(interval);
-
-            if ( cacheHit ) {
-                ++numCacheHits;
-            }
-            else {
-                ++numCacheMisses;
-            }
-
-            return cacheHit;
-        }
-
-        /**
-         * Trims the cache to the specified new start position by discarding all records that end before it
-         * while preserving relative ordering of records.
-         *
-         * @param newStart new start position on the current contig to which to trim the cache
-         */
-        public void trimToNewStartPosition( final int newStart ) {
-            if ( newStart > cachedInterval.getEnd() ) {
-                throw new GATKException(String.format("BUG: attempted to trim Feature cache to an improper new start position (%d). Cache stop = %d",
-                                                      newStart, cachedInterval.getEnd()));
-            }
-
-            List<CACHED_FEATURE> overlappingFeaturesBeforeNewStart = new ArrayList<>(EXPECTED_MAX_OVERLAPPING_FEATURES_DURING_CACHE_TRIM);
-
-            // In order to trim the cache to the new start position, we need to find
-            // all Features in the cache that start before the new start position,
-            // and discard those that don't overlap the new start while keeping those
-            // that do overlap. We can stop once we find a Feature that starts on or
-            // after the new start position, since the Features are assumed to be sorted
-            // by start position.
-            while ( ! cache.isEmpty() && cache.getFirst().getStart() < newStart ) {
-                CACHED_FEATURE featureBeforeNewStart = cache.removeFirst();
-
-                if ( featureBeforeNewStart.getEnd() >= newStart ) {
-                    overlappingFeaturesBeforeNewStart.add(featureBeforeNewStart);
-                }
-            }
-
-            // Add back the Features that started before the new start but overlapped it
-            // in the reverse of the order in which we encountered them so that their original
-            // relative ordering in the cache is restored.
-            for ( int i = overlappingFeaturesBeforeNewStart.size() - 1; i >= 0; --i ) {
-                cache.addFirst(overlappingFeaturesBeforeNewStart.get(i));
-            }
-
-            // Record our new start boundary
-            cachedInterval = new SimpleInterval(cachedInterval.getContig(), newStart, cachedInterval.getEnd());
-        }
-
-        /**
-         * Returns (but does not remove) all cached Features that overlap the region from the start
-         * of our cache (cacheStart) to the specified stop position.
-         *
-         * @param stopPosition Endpoint of the interval that returned Features must overlap
-         * @return all cached Features that overlap the region from the start of our cache to the specified stop position
-         */
-        public List<CACHED_FEATURE> getCachedFeaturesUpToStopPosition( final int stopPosition ) {
-            List<CACHED_FEATURE> matchingFeatures = new ArrayList<>(cache.size());
-
-            // Find (but do not remove from our cache) all Features that start before or on the provided stop position
-            for ( CACHED_FEATURE candidateFeature : cache ) {
-                if ( candidateFeature.getStart() > stopPosition ) {
-                    break; // No more possible matches among the remaining cached Features, so stop looking
-                }
-                matchingFeatures.add(candidateFeature);
-            }
-            return matchingFeatures;
-        }
-    }
-
-    /**
      * Creates a FeatureDataSource backed by the provided file that uses the given codec to decode records
      * from that file. The data source will have no name, and will look ahead the default number of bases
      * ({@link #DEFAULT_QUERY_LOOKAHEAD_BASES}) during queries that produce cache misses.
@@ -450,7 +237,7 @@ public final class FeatureDataSource<T extends Feature> implements GATKDataSourc
 
         try {
             // Save the iterator returned so that we can close it properly later
-            currentIterator = intervalsForTraversal != null ? new FeatureIntervalIterator(intervalsForTraversal)
+            currentIterator = intervalsForTraversal != null ? new FeatureIntervalIterator<T>(intervalsForTraversal, featureReader, featureFile.getAbsolutePath())
                                                             : featureReader.iterator();
             return currentIterator;
         }
@@ -595,7 +382,9 @@ public final class FeatureDataSource<T extends Feature> implements GATKDataSourc
     @Override
     public void close() {
         closeOpenIterationIfNecessary();
-        printCacheStatistics();
+
+        logger.debug(String.format("Cache statistics for FeatureInput %s:", name != null ? name + " (" + featureFile.getName() + ")" : featureFile.getName()));
+        queryCache.printCacheStatistics();
 
         try {
             if ( featureReader != null )
@@ -614,154 +403,5 @@ public final class FeatureDataSource<T extends Feature> implements GATKDataSourc
             currentIterator.close();
             currentIterator = null;
         }
-    }
-
-    /**
-     * Print statistics about the cache hit rate for debugging
-     */
-    private void printCacheStatistics() {
-        final int totalQueries = queryCache.getNumCacheHits() + queryCache.getNumCacheMisses();
-        logger.debug(String.format("For feature input %s: Cache hit rate was %.2f%% (%d out of %d total queries)",
-                     name != null ? name + " (" + featureFile.getName() + ")" : featureFile.getName(),
-                     totalQueries > 0 ? ((double)queryCache.getNumCacheHits() / totalQueries) * 100.0 : 0.0,
-                     queryCache.getNumCacheHits(),
-                     totalQueries));
-    }
-
-    /**
-     * Iterator implementation of Feature traversal by intervals.
-     *
-     * Given a List of intervals, queries our backing data source for Features overlapping
-     * each successive interval and iterates over them, while also guaranteeing that each
-     * Feature overlapping our intervals will only be returned once during the entire iteration.
-     *
-     * Requires that the provided List of intervals consist of non-overlapping intervals
-     * sorted in increasing order of start position.
-     */
-    private class FeatureIntervalIterator implements CloseableTribbleIterator<T> {
-        private final Iterator<SimpleInterval> intervalIterator;
-        private CloseableTribbleIterator<T> featuresInCurrentInterval;
-        private T nextFeature;
-        private SimpleInterval currentInterval;
-        private SimpleInterval previousInterval;
-
-        /**
-         * Initialize a FeatureIntervalIterator with a set of intervals.
-         *
-         * Requires that the provided List of intervals consist of non-overlapping intervals
-         * sorted in increasing order of start position.
-         *
-         * @param intervals intervals to use for traversal. must be non-overlapping and sorted by increasing start position.
-         */
-        public FeatureIntervalIterator( final List<SimpleInterval> intervals ) {
-            this.intervalIterator = intervals.iterator();
-            nextFeature = loadNextNovelFeature();
-        }
-
-        @Override
-        public boolean hasNext() {
-            return nextFeature != null;
-        }
-
-        @Override
-        public T next() {
-            if ( nextFeature == null ) {
-                throw new NoSuchElementException("No more Features for current interval set");
-            }
-
-            final T toReturn = nextFeature;
-            nextFeature = loadNextNovelFeature();
-            return toReturn;
-        }
-
-        /**
-         * @return the next Feature from our data source that we HAVEN'T previously encountered,
-         *         or null if we're out of Features
-         */
-        private T loadNextNovelFeature() {
-            T candidateFeature;
-
-            do {
-                candidateFeature = loadNextFeature();
-
-                if ( candidateFeature != null && featureIsNovel(candidateFeature) ) {
-                    return candidateFeature;
-                }
-            } while ( candidateFeature != null );
-
-            return null;
-        }
-
-        /**
-         * @return the next Feature from our data source (regardless of whether we've encountered it before or not),
-         *         or null if we're out of Features
-         */
-        private T loadNextFeature() {
-            // If we're out of Features for the current interval, repeatedly query the next interval
-            // until we find one with overlapping Features. Return null if we run out of intervals.
-            while ( featuresInCurrentInterval == null || ! featuresInCurrentInterval.hasNext() ) {
-                if ( ! queryNextInterval() ) {
-                    return null;
-                }
-            }
-
-            // If we reach here, we're guaranteed to have at least one Feature left to consume in the current interval.
-            return featuresInCurrentInterval.next();
-        }
-
-        /**
-         * Determines whether a Feature is novel (hasn't been encountered before on this iteration). A Feature
-         * hasn't been encountered before if we're either on the very first query interval, or the Feature doesn't
-         * overlap our previous query interval.
-         *
-         * @param feature Feature to test
-         * @return true if we haven't seen the Feature before on this iteration, otherwise false
-         */
-        private boolean featureIsNovel( final T feature ) {
-            return previousInterval == null || ! previousInterval.overlaps(new SimpleInterval(feature));
-        }
-
-        /**
-         * Performs a query on the next interval in our interval List, and initializes all members appropriately
-         * to prepare for processing the Features overlapping that interval.
-         *
-         * @return true if we successfully queried the next interval, false if there are no more intervals to query
-         */
-        private boolean queryNextInterval() {
-            // Make sure to close out the query iterator for the previous interval, since Tribble only allows us
-            // to have one iterator open over our FeatureReader at a time.
-            if ( featuresInCurrentInterval != null ) {
-                featuresInCurrentInterval.close();
-                featuresInCurrentInterval = null;
-            }
-
-            if ( ! intervalIterator.hasNext() ) {
-                currentInterval = previousInterval = null;
-                return false;
-            }
-
-            previousInterval = currentInterval;
-            currentInterval = intervalIterator.next();
-            try {
-                featuresInCurrentInterval = featureReader.query(currentInterval.getContig(), currentInterval.getStart(), currentInterval.getEnd());
-                return true;
-            }
-            catch ( IOException e ) {
-                throw new GATKException("Error querying file " + featureFile.getAbsolutePath() + " over interval " + currentInterval, e);
-            }
-        }
-
-        @Override
-        public Iterator<T> iterator() {
-            return this;
-        }
-
-        @Override
-        public void close() {
-            if ( featuresInCurrentInterval != null ) {
-                featuresInCurrentInterval.close();
-            }
-        }
-
     }
 }
